Proper logging for the bot.
This is important because the bot has been
moved to a remote server and it would be
impossible to debug issues without logs.

The logs should take up less than 15MB of
space in total, and I estimate that they should
account for about 15 days of logs, assuming that
the bot runs every 3 minutes (it does as of now).